+++ servo/servo.c
@@ -8,7 +8,6 @@
 *
 *******************************************************************************/
 
-
 /*------------------------------------------------------------------------------
  Standard Includes                                                                     
 ------------------------------------------------------------------------------*/
@@ -23,15 +22,11 @@
 #include "led.h"
 #include "init.h"
 #include "usb.h"
-
-/*------------------------------------------------------------------------------
- Literal Constants
-------------------------------------------------------------------------------*/
-#define SER_PER 0.55555
-
 /*------------------------------------------------------------------------------
  Global Variables 
 ------------------------------------------------------------------------------*/
+#define SER_PER 0.55555
+
 extern SERVO_PRESET servo_preset;
 
 /*------------------------------------------------------------------------------
@@ -44,42 +39,72 @@ extern SERVO_PRESET servo_preset;
 * 		servo_init                                                             *
 *                                                                              *
 * DESCRIPTION:                                                                 * 
-* 		Initialize/Reset servo                                                 *
+* 		Initialize/Reset servo                             *
 *                                                                              *
 *******************************************************************************/
-SERVO_STATUS servo_init
-    (
-    void
-    )
+SERVO_STATUS servo_init()
 {
-/* GPIO Initialization */
-HAL_GPIO_WritePin(MOTOR1_EN_PORT, MOTOR1_EN, GPIO_PIN_SET);
-HAL_GPIO_WritePin(MOTOR2_EN_PORT, MOTOR2_EN, GPIO_PIN_SET);
-HAL_GPIO_WritePin(MOTOR3_EN_PORT, MOTOR3_EN, GPIO_PIN_SET);
-HAL_GPIO_WritePin(MOTOR4_EN_PORT, MOTOR4_EN, GPIO_PIN_SET);
+    // GPIO Initialization
+    HAL_GPIO_WritePin(MOTOR1_EN_PORT, MOTOR1_EN, GPIO_PIN_SET);
+    HAL_GPIO_WritePin(MOTOR2_EN_PORT, MOTOR2_EN, GPIO_PIN_SET);
+    HAL_GPIO_WritePin(MOTOR3_EN_PORT, MOTOR3_EN, GPIO_PIN_SET);
+    HAL_GPIO_WritePin(MOTOR4_EN_PORT, MOTOR4_EN, GPIO_PIN_SET);
 
-/* Timer intialization */
-HAL_StatusTypeDef hal_status1, hal_status2, hal_status3, hal_status4;
+    // Timer intialization
+    HAL_StatusTypeDef hal_status1, hal_status2, hal_status3, hal_status4;
 
-hal_status1 = HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_4);
-hal_status2 = HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_3);
-hal_status3 = HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_1);
-hal_status4 = HAL_TIM_PWM_Start(&htim2, TIM_CHANNEL_1);
+    hal_status1 = HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_4);
+    hal_status2 = HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_3);
+    hal_status3 = HAL_TIM_PWM_Start(&htim3, TIM_CHANNEL_1);
+    hal_status4 = HAL_TIM_PWM_Start(&htim2, TIM_CHANNEL_1);
 
-if ( hal_status1 == HAL_OK &&
-    hal_status2 == HAL_OK &&
-    hal_status3 == HAL_OK &&
-    hal_status4 == HAL_OK )
-    {
-    return SERVO_OK;
-    }
-else
-    {
-    return SERVO_FAIL;
-    }
-
+    if ( hal_status1 == HAL_OK &&
+        hal_status2 == HAL_OK &&
+        hal_status3 == HAL_OK &&
+        hal_status4 == HAL_OK )
+        {
+        return SERVO_OK;
+        }
+        else
+        {
+        return SERVO_FAIL;
+        }
 }
 
+/*******************************************************************************
+*                                                                              *
+* PROCEDURE:                                                                   * 
+* 		motor_drive                                                            *
+*                                                                              *
+* DESCRIPTION:                                                                 * 
+* 		Drives the specified servo motor to the desired angle (0-180).         *
+*                                                                              *
+* NOTES:                                                                       *
+* 		- Replaces previous motor1_drive, motor2_drive, etc.                   *
+* 		- Uses ServoID enum to select which servo to control.                  *
+*                                                                              *
+*******************************************************************************/
+void motor_drive(ServoID servo, uint8_t angle)
+{
+    uint8_t pulse = angle_to_pulse(motor_snap_to_bound(angle, 180, 0));
+    switch (servo) {
+        case SERVO_1:
+            htim3.Instance->CCR4 = pulse;
+            break;
+        case SERVO_2:
+            htim3.Instance->CCR3 = pulse;
+            break;
+        case SERVO_3:
+            htim3.Instance->CCR1 = pulse;
+            break;
+        case SERVO_4:
+            htim2.Instance->CCR1 = pulse;
+            break;
+        default:
+            // Optionally: handle error
+            break;
+    }
+}
 
 /*******************************************************************************
 *                                                                              *
@@ -87,89 +112,16 @@ else
 * 		servo_reset                                                            *
 *                                                                              *
 * DESCRIPTION:                                                                 * 
-* 		Initialize/Reset servo rotate to default                               *
+* 		Resets all servos to their preset positions                            *
 *                                                                              *
 *******************************************************************************/
-void servo_reset
-    (
-    void
-    )
+void servo_reset()
 {
-    motor1_drive(servo_preset.rp_servo1);
-    motor2_drive(servo_preset.rp_servo2);
-    motor3_drive(servo_preset.rp_servo3);
-    motor4_drive(servo_preset.rp_servo4);
-
+    motor_drive(SERVO_1, servo_preset.rp_servo1);
+    motor_drive(SERVO_2, servo_preset.rp_servo2);
+    motor_drive(SERVO_3, servo_preset.rp_servo3);
+    motor_drive(SERVO_4, servo_preset.rp_servo4);
 }
-
-
-/*******************************************************************************
-*                                                                              *
-* PROCEDURE:                                                                   * 
-* 		motor1_drive                                                           *
-*                                                                              *
-* DESCRIPTION:                                                                 * 
-* 		Drive the first servo motor with a desired angle (0-180)               *
-*                                                                              *
-*******************************************************************************/
-void motor1_drive(uint8_t angle)
-{
-    uint8_t pulse = angle_to_pulse(motor_snap_to_bound(angle, 180, 0));
-    htim3.Instance->CCR4 = pulse;
-
-}
-
-
-/*******************************************************************************
-*                                                                              *
-* PROCEDURE:                                                                   * 
-* 		motor2_drive                                                           *
-*                                                                              *
-* DESCRIPTION:                                                                 * 
-* 		Drive the second servo motor with a desired angle (0-180)              *
-*                                                                              *
-*******************************************************************************/
-void motor2_drive(uint8_t angle)
-{
-    uint8_t pulse = angle_to_pulse(motor_snap_to_bound(angle, 180, 0));
-    htim3.Instance->CCR3 = pulse;
-
-}
-
-
-/*******************************************************************************
-*                                                                              *
-* PROCEDURE:                                                                   * 
-* 		motor3_drive                                                           *
-*                                                                              *
-* DESCRIPTION:                                                                 * 
-* 		Drive the third servo motor with a desired angle (0-180)               *
-*                                                                              *
-*******************************************************************************/
-void motor3_drive(uint8_t angle)
-{
-    uint8_t pulse = angle_to_pulse(motor_snap_to_bound(angle, 180, 0));
-    htim3.Instance->CCR1 = pulse;
-
-}
-
-
-/*******************************************************************************
-*                                                                              *
-* PROCEDURE:                                                                   * 
-* 		motor4_drive                                                           *
-*                                                                              *
-* DESCRIPTION:                                                                 * 
-* 		Drive the forth servo motor with a desired angle (0-180)               *
-*                                                                              *
-*******************************************************************************/
-void motor4_drive(uint8_t angle)
-{
-    uint8_t pulse = angle_to_pulse(motor_snap_to_bound(angle, 180, 0));
-    htim2.Instance->CCR1 = pulse;
-
-}
-
 
 /*******************************************************************************
 *                                                                              *
@@ -178,18 +130,16 @@ void motor4_drive(uint8_t angle)
 *                                                                              *
 * DESCRIPTION:                                                                 * 
 * 		A complete function that drives all servos in this board               *
+*       Uses motor_drive for each servo.                                       *
 *                                                                              *
 *******************************************************************************/
-void motors_drive
-    (
-    uint8_t angle
-    )
-{
-    motor1_drive(angle);
-    motor2_drive(angle);
-    motor3_drive(angle);
-    motor4_drive(angle);
 
+void motors_drive(uint8_t angle)
+{
+    motor_drive(SERVO_1, angle);
+    motor_drive(SERVO_2, angle);
+    motor_drive(SERVO_3, angle);
+    motor_drive(SERVO_4, angle);
 }
 
 
@@ -202,21 +152,16 @@ void motors_drive
 * 		Executes a servo command given from the terminal.                      *
 *                                                                              *
 *******************************************************************************/
-SERVO_STATUS servo_cmd_execute
-    (
-    uint8_t subcommand
-    )
-{
-USB_STATUS usb_status;
+SERVO_STATUS servo_cmd_execute(uint8_t subcommand){
+    USB_STATUS usb_status;
 
-switch ( subcommand )
-    {
-    case SERVO_SWEEP:
+    switch (subcommand){
+        case SERVO_SWEEP:
         {
             uint8_t degree;
             usb_status = usb_receive(&degree, sizeof(uint8_t), 1000);
             motors_drive(degree);
-            if ( usb_status == USB_OK ) 
+            if (usb_status == USB_OK) 
                 {
                 return SERVO_OK;
                 }
@@ -235,9 +180,7 @@ switch ( subcommand )
         default:
             return SERVO_FAIL;
     }
-
 }
-
 
 /*******************************************************************************
 *                                                                              *
@@ -251,38 +194,15 @@ switch ( subcommand )
 uint8_t angle_to_pulse(uint8_t angle)
 {
     return 25 + (angle*SER_PER);
-
 }
 
-
-/*******************************************************************************
-*                                                                              *
-* PROCEDURE:                                                                   * 
-* 		motor_snap_to_bound                                                    *
-*                                                                              *
-* DESCRIPTION:                                                                 * 
-* 		Keep an angle within the specified actuation range.                    *
-*                                                                              *
-*******************************************************************************/
-uint8_t motor_snap_to_bound
-    (
-    uint8_t angle, 
-    uint8_t upper, 
-    uint8_t lower
-    )
+uint8_t motor_snap_to_bound(uint8_t angle, uint8_t upper, uint8_t lower)
 {
-if (angle >= lower && angle <= upper) 
-    {
-    return angle;
-    } 
-else if ( angle > upper 
-       && angle <= ( upper + ( ( 255 - upper ) / 2 ) ) ) 
-    {
-    return upper;
-    } 
-else 
-    {
-    return lower;
+    if (angle >= lower && angle <= upper) {
+        return angle;
+    } else if (angle > upper && angle <= (upper + ((255 - upper) / 2))) {
+        return upper;
+    } else {
+        return lower;
     }
-
 }
